# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: REST API for managing a library.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /api/loans:
    get:
      tags:
        - Loans
      summary: Get all loans
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanResponse'

    post:
      tags:
        - Loans
      summary: Borrow a book
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanRequest'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
  /api/loans/{id}:
    get:
      tags:
        - Loans
      summary: Get loan by id
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Loan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'

        '404':
          description: Loan not found

    delete:
      tags:
        - Loans
      summary: Delete a loan
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Loan deleted
  /api/loans/{id}/return:
    put:
      tags:
        - Loans
      summary: Return a book
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'

        '404':
          description: Loan not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoanRequest:
      type: object
      required:
        - userId
        - bookId

      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
          example: 1

        bookId:
          type: integer
          format: int64
          minimum: 1
          example: 2
    LoanResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        userName:
          type: string
          example: Sabrina Moufok

        bookTitle:
          type: string
          example: Clean Code

        BorrowDate:
          type: string
          format: date
          example: "2026-07-27"

        returnDate:
          type: string
          format: date
          example: "2026-08-10"

        loanStatus:
          type: string
          example: BORROWED